+++ common/src/main/java/com/reazip/economycraft/EconomySources.java
@@ -7,6 +7,7 @@ public final class EconomySources {
     public static final MutationSource ADMIN_ADD = MutationSource.of("economycraft:admin_add");
     public static final MutationSource ADMIN_REMOVE = MutationSource.of("economycraft:admin_remove");
     public static final MutationSource ADMIN_SET = MutationSource.of("economycraft:admin_set");
+    public static final MutationSource ADMIN_RESET = MutationSource.of("economycraft:admin_reset");
     public static final MutationSource DAILY_REWARD = MutationSource.of("economycraft:daily_reward");
     public static final MutationSource SHOP_PURCHASE = MutationSource.of("economycraft:shop_purchase");
     public static final MutationSource SHOP_SALE = MutationSource.of("economycraft:shop_sale");